refactor: fold the rate limiter logic file into the component
A component's files are component.ts, types.ts, index.ts and an optional
errors.ts. The extra logic.ts followed the shape of the
http-requests-logger and http-tracer components rather than the
standard, so the pure helpers move into component.ts alongside the
factory that uses them, and the two spec files collapse into one now
that there is a single source file to cover.

No behaviour or public surface change: the helpers were already
re-exported through index.ts and still are.
