+++ components/rate-limiter/src/index.ts
@@ -1,4 +1,3 @@
 export * from './types'
 export * from './errors'
-export * from './logic'
 export * from './component'
